Added systemd service support.
Rejiggered the file locations to cleanly allow for other specfiles to live in this repo.
Added the systemd unit files to live next to the specfile itself.
Now to recreate the rpm file, we need to create the source, copy the spec files into the source folder, and then create the final source tar.
It's a little bit of manual work, but I hope to not have to do this often once everything is stable.