Hardening the unit files.
They all now depend on network-online.target which will prevent things like namenode listening on a localhost ip.
Added a hack for the MR History Server that has systemd restart the service whenever it goes down.
If HDFS is not available at the time of starting the MR History Server, then it will immediately fail instead of re-attempting to connect.