Changing every RPM to install to /opt.
It is becoming too much of a pain to make every process FHS compliant.
I want to be able to easily add and upgrade processes without remembering how I got them to play nice with FHS.
The simple and dumb /opt model allows for weirdo processes *cough* solr and pig *cough* to be easily maintained on my side.